requirements.txt
django>=1.11.<1.12
selenium>=3.4.1,<4.0
django_leaflet>=0.22,<0.23
django-bootstrap3>=8.2.3,<9.0
paho-mqtt>=1.2.3,<2.0